requirements.txt
Django==4.0.2
django-crispy-forms
Pillow
